requirements.txt
pyrogram
pyromod
hachoir
tgcrypto
uvloop
requests
aiohttp
aiofiles
python-dotenv
firebase-admin